refactor: Extract EMIT methane layer to dedicated component - Extract methane polygon and PNG rendering for SyncMapTracking - Create dedicated MethaneEMITLayer component with proper props flow - Maintain functionaliy in mapViewB while improving code separation
